backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.34.0
python-multipart==0.0.17
pydantic==2.10.3
pydantic-settings==2.6.1
python-dotenv==1.0.1
google-genai
gspread==6.1.4
oauth2client==4.1.3
pdf2image==1.17.0
pypdf2==3.0.1
pillow==11.0.0
fuzzywuzzy==0.18.0
python-Levenshtein==0.25.0
pandas==2.2.0
